+++ src/main/kotlin/io/extremum/functions/api/function/model/StorageTriggerParameters.kt
@@ -7,9 +7,8 @@ data class StorageTriggerParameters(
 ) {
     data class Instance(
         val id: String,
-        val type: String = DEFAULT_TYPE,
+        val type: String,
         val created: Date,
-        val operation: String,
         val key: String,
         val meta: Meta,
     )
@@ -17,8 +16,4 @@ data class StorageTriggerParameters(
     data class Meta(
         val trace: String,
     )
-
-    private companion object {
-        const val DEFAULT_TYPE = "StorageEvent"
-    }
 }